new errorRMSmap function, estimates error on parameters
This new function will take in entry a list of folders. In each folder it
will look for the files of the RMS map and the file of the fitted data.
It will extrapolate the RMS map and calculate the gradients and second
order crossed gradients as well as the error along each parameters.